requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alabaster==1.0.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astroid==3.3.11
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
asyncssh==2.21.1
atomicwrites==1.4.1
attrs==25.4.0
autopep8==2.0.4
babel==2.17.0
backcall==0.2.0
backports.tarfile==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
binaryornot==0.4.4
black==25.11.0
bleach==6.3.0
boto3==1.41.4
botocore==1.41.4
cattrs==25.3.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
clarabel==0.11.1
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cookiecutter==2.6.0
cramjam==2.11.0
cryptography==46.0.3
cvxpy==1.7.4
cycler==0.12.1
DateTime==6.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
diff-match-patch==20241021
dill==0.4.0
docopt==0.6.2
docstring-to-markdown==0.17
docutils==0.21.2
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
fastparquet==2024.11.0
filelock==3.20.0
flake8==7.1.2
flexcache==0.3
flexparser==0.4
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gurobipy==13.0.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imagesize==1.4.1
importlib_metadata==8.7.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.31.0
ipython==8.12.3
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isort==6.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
jellyfish==1.2.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keyring==25.7.0
kiwisolver==1.4.9
lark==1.3.1
llvmlite==0.45.1
lsprotocol==2025.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.5.0
notebook_shim==0.2.4
NREL-PySAM==6.0.1
numba==0.62.1
numpy==1.26.4
numpydoc==1.9.0
ochre-nrel==0.9.1
osqp==1.0.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.0.0
Pint==0.24.4
pip==25.3
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
PsychroLib==2.5.0
ptyprocess==0.7.0
pure_eval==0.2.3
pvlib==0.13.1
pyarrow==22.0.0
pycodestyle==2.12.1
pycparser==2.23
pydocstyle==6.3.0
pyflakes==3.2.0
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.9
pylint-venv==3.0.4
pyls-spyder==0.4.0
PyNaCl==1.6.1
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-ruff==2.3.0
python-lsp-server==1.13.2
python-slugify==8.0.4
pytokens==0.3.0
pytoolconfig==1.3.1
pytz==2024.2
pyuca==1.2
pywin32-ctypes==0.2.3
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
QDarkStyle==3.2.3
qstylizer==0.2.4
QtAwesome==1.4.0
qtconsole==5.7.0
QtPy==2.4.3
rainflow==3.2.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rope==1.14.0
rpds-py==0.29.0
rtree==1.4.1
ruff==0.14.6
s3transfer==0.15.0
scikit-learn==1.7.2
scipy==1.15.3
scs==3.2.9
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.8
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==6.1.1
spyder-kernels==3.1.2
stack_data==0.6.3
superqt==0.7.6
sympy==1.14.0
terminado==0.18.1
text-unidecode==1.3
textdistance==4.6.3
threadpoolctl==3.6.0
three-merge==0.1.1
tinycss2==1.4.0
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.1
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uri-template==1.3.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
whatthepatch==1.0.7
wheel==0.45.1
xmltodict==0.15.1
yapf==0.43.0
yarg==0.1.9
yarl==1.22.0
zipp==3.23.0
zope.interface==8.1.1
